+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/InvoiceEditPanel.java
@@ -51,7 +51,6 @@ public final class InvoiceEditPanel extends EntityEditPanel {
     super(editModel);
     this.invoiceLinePanel = invoiceLinePanel;
     clearAfterInsert().set(false);
-    defaults().textFieldColumns().set(12);
   }
 
   @Override
@@ -65,6 +64,7 @@ public final class InvoiceEditPanel extends EntityEditPanel {
             .columns(6);
 
     createTextField(Invoice.BILLINGADDRESS)
+            .columns(12)
             .selectAllOnFocusGained(true);
 
     createTextField(Invoice.BILLINGCITY)
@@ -125,7 +125,7 @@ public final class InvoiceEditPanel extends EntityEditPanel {
       tableModel.columnModel().setVisibleColumns(Customer.LASTNAME, Customer.FIRSTNAME, Customer.EMAIL);
       tableModel.sortModel().setSortOrder(Customer.LASTNAME, ASCENDING);
       tableModel.sortModel().addSortOrder(Customer.FIRSTNAME, ASCENDING);
-      selector.setPreferredSize(new Dimension(500, 300));
+      selector.preferredSize(new Dimension(500, 300));
 
       return selector;
     }