+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/icon/DefaultFontImageIcon.java
@@ -48,7 +48,7 @@ final class DefaultFontImageIcon implements FontImageIcon {
   }
 
   @Override
-  public void setColor(Color color) {
+  public void color(Color color) {
     fontIcon.setIconColor(color);
     iconPainter.paintIcon(fontIcon, imageIcon);
   }