+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/icon/DefaultIcons.java
@@ -65,9 +65,9 @@ final class DefaultIcons implements Icons {
   }
 
   @Override
-  public void setIconColor(Color color) {
+  public void iconColor(Color color) {
     requireNonNull(color);
-    icons.values().forEach(icon -> icon.setColor(color));
+    icons.values().forEach(icon -> icon.color(color));
   }
 
   @Override
@@ -87,7 +87,7 @@ final class DefaultIcons implements Icons {
     @Override
     public void accept(Color color) {
       if (color != null) {
-        setIconColor(color);
+        iconColor(color);
       }
     }
   }